backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
supabase==2.0.0
pydantic==2.5.0
spacy==3.7.2
pypdf>=3.17.1
scikit-learn==1.3.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
httpx==0.25.2
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
slowapi==0.1.9
pydantic-settings==2.1.0
requests==2.31.0
numpy>=1.24.0
pandas>=2.0.0